@@ -14,32 +14,32 @@ Future<void> _firebaseMessagingBackgroundHandler(RemoteMessage message) async {
 
 class AppStateModel with ChangeNotifier {
   bool _authenticated = false;
-  String _userToken;
-  String _idToken;
-  String _refreshToken;
-  DateTime _expires;
-  String _email;
-  String _name;
-  String _locale;
-  String _fcmToken;
-  SharedPreferences prefs;
-  FirebaseAnalytics analytics;
-  FirebaseMessaging messaging;
+  String? _userToken;
+  String? _idToken;
+  String? _refreshToken;
+  DateTime? _expires;
+  String? _email;
+  String? _name;
+  String? _locale;
+  String? _fcmToken;
+  SharedPreferences? prefs;
+  FirebaseAnalytics? analytics;
+  FirebaseMessaging? messaging;
   // We use a mockmap to enable and disable mock functions/classes.
   // The mock should be injected as a dependency where external dependencies need
   // to be mocked as part of testing.
   MockMap _mocks = MockMap();
 
   bool get authenticated => _authenticated;
-  String get userToken => _userToken;
-  String get idToken => _idToken;
-  String get refreshToken => _refreshToken;
-  DateTime get expires => _expires;
-  String get email => _email;
-  String get name => _name;
+  String? get userToken => _userToken;
+  String? get idToken => _idToken;
+  String? get refreshToken => _refreshToken;
+  DateTime? get expires => _expires;
+  String? get email => _email;
+  String? get name => _name;
   MockMap get mocks => _mocks;
-  String get locale => _locale;
-  String get fcmToken => _fcmToken;
+  String? get locale => _locale;
+  String? get fcmToken => _fcmToken;
 
   AppStateModel([this.prefs, this.analytics, this.messaging]) {
     refresh();
@@ -66,7 +66,7 @@ class AppStateModel with ChangeNotifier {
       _fcmToken = 'only_available_in_js';
       return;
     }
-    NotificationSettings settings = await messaging.requestPermission(
+    NotificationSettings settings = await messaging!.requestPermission(
       alert: true,
       announcement: false,
       badge: true,
@@ -86,29 +86,29 @@ class AppStateModel with ChangeNotifier {
 
     FirebaseMessaging.onBackgroundMessage(_firebaseMessagingBackgroundHandler);
 
-    messaging.getToken().then((String token) {
+    messaging!.getToken().then((String? token) {
       assert(token != null);
       print("Firebase messaging token: $token");
       _fcmToken = token;
     });
-    messaging.onTokenRefresh.listen((newToken) {
+    messaging!.onTokenRefresh.listen((newToken) {
       _fcmToken = newToken;
     });
   }
 
-  void setLocale(String loc) {
+  void setLocale(String? loc) {
     if (prefs == null) {
       return;
     }
     if (loc == null) {
-      loc = prefs.getString('locale');
+      loc = prefs!.getString('locale');
       if (loc == null) {
         loc = Intl.getCurrentLocale().substring(0, 2);
       }
     }
     _locale = loc;
-    S.load(Locale(_locale));
-    prefs.setString('locale', _locale);
+    S.load(Locale(_locale!));
+    prefs!.setString('locale', _locale!);
     notifyListeners();
   }
 
@@ -130,11 +130,11 @@ class AppStateModel with ChangeNotifier {
   }
 
   Future<void> sendAnalyticsEvent(
-      String name, Map<String, dynamic> params) async {
+      String name, Map<String, dynamic>? params) async {
     if (this.analytics == null) {
       return;
     }
-    await this.analytics.logEvent(
+    await this.analytics!.logEvent(
           name: name,
           parameters: params,
         );
@@ -146,19 +146,19 @@ class AppStateModel with ChangeNotifier {
       return;
     }
     // Check if the stored token has expired
-    var expiresStr = prefs.getString('expires');
+    var expiresStr = prefs!.getString('expires');
     if (expiresStr != null) {
       _expires = DateTime.parse(expiresStr);
-      var remaining = _expires.difference(DateTime.now());
+      var remaining = _expires!.difference(DateTime.now());
       if (remaining.inSeconds < 3600) {
-        Map<dynamic, dynamic> auth;
-        auth = await AuthClient(authClient: _mocks.getMock('authClient'))
-            .refreshToken(prefs.getString('refreshToken'));
+        Map<dynamic, dynamic>? auth;
+        auth = await AuthClient(authClient: _mocks.getAppAuth())
+            .refreshToken(prefs!.getString('refreshToken'));
         if (auth != null && auth.containsKey('access_token')) {
           logIn(auth);
         } else {
-          prefs.remove('userToken');
-          prefs.remove('expires');
+          prefs!.remove('userToken');
+          prefs!.remove('expires');
           _authenticated = false;
           _userToken = null;
           _expires = null;
@@ -167,12 +167,12 @@ class AppStateModel with ChangeNotifier {
         return;
       }
     }
-    _userToken = prefs.getString('userToken');
+    _userToken = prefs!.getString('userToken');
     if (_userToken != null) {
       _authenticated = true;
     }
-    _idToken = prefs.getString('idToken');
-    _refreshToken = prefs.getString('refreshToken');
+    _idToken = prefs!.getString('idToken');
+    _refreshToken = prefs!.getString('refreshToken');
     notifyListeners();
   }
 
@@ -181,11 +181,11 @@ class AppStateModel with ChangeNotifier {
       return;
     }
     if (data.containsKey('email')) {
-      prefs.setString('email', data['email']);
+      prefs!.setString('email', data['email']);
       _email = data['email'];
     }
     if (data.containsKey('name')) {
-      prefs.setString('name', data['name']);
+      prefs!.setString('name', data['name']);
       _name = data['name'];
     }
     notifyListeners();
@@ -196,21 +196,21 @@ class AppStateModel with ChangeNotifier {
       return;
     }
     if (data.containsKey('access_token')) {
-      prefs.setString('userToken', data['access_token']);
+      prefs!.setString('userToken', data['access_token']);
       _userToken = data['access_token'];
       _authenticated = true;
     }
     if (data.containsKey('refresh_token')) {
-      prefs.setString('refreshToken', data['refresh_token']);
+      prefs!.setString('refreshToken', data['refresh_token']);
       _refreshToken = data['refresh_token'];
     }
     if (data.containsKey('id_token')) {
-      prefs.setString('idToken', data['id_token']);
+      prefs!.setString('idToken', data['id_token']);
       _idToken = data['id_token'];
     }
     if (data.containsKey('expires')) {
       _expires = data['expires'];
-      prefs.setString('expires', _expires.toIso8601String());
+      prefs!.setString('expires', _expires!.toIso8601String());
     }
     sendAnalyticsEvent('login', null);
     notifyListeners();
@@ -221,13 +221,13 @@ class AppStateModel with ChangeNotifier {
        (if supported). closeSessions() is not implemented here as it
        involves clearing cookies in the webview (for demo.identityprovider.io).
     */
-    //AuthClient(authClient:_mocks.getMock('authClient')).closeSessions();
+    //AuthClient(authClient:_mocks.getAppAuth().closeSessions();
     _authenticated = false;
     _userToken = null;
     _idToken = null;
     _refreshToken = null;
     _expires = null;
-    prefs.clear();
+    prefs!.clear();
     notifyListeners();
   }
 }
