@@ -11,7 +11,7 @@ class Geo {
 
   Future<List<Placemark>> placemarkFromCoordinates(
       double latitude, double longitude,
-      {String localeIdentifier}) {
+      {String? localeIdentifier}) {
     return placemarkFromCoordinates(latitude, longitude);
   }
 
@@ -19,8 +19,8 @@ class Geo {
       {LocationAccuracy desiredAccuracy = LocationAccuracy.best,
       int distanceFilter = 0,
       bool forceAndroidLocationManager = false,
-      Duration intervalDuration,
-      Duration timeLimit}) {
+      Duration? intervalDuration,
+      Duration? timeLimit}) {
     return Geolocator.getPositionStream(
         desiredAccuracy: desiredAccuracy,
         distanceFilter: distanceFilter,
@@ -31,7 +31,7 @@ class Geo {
 }
 
 class LocStateModel with ChangeNotifier {
-  Geo locator;
+  Geo? locator;
   // Represent latest location
   Position _lastPos = Position(
       longitude: 0.0,
@@ -42,9 +42,10 @@ class LocStateModel with ChangeNotifier {
       altitude: 0.0,
       heading: 0.0,
       timestamp: DateTime.now());
-  LocationPermission _geoAccessStatus;
-  StreamSubscription<Position> _positionStreamSubscription;
-  final Map<Position, Placemark> _pointList = Map<Position, Placemark>();
+  LocationPermission? _geoAccessStatus;
+  // ignore: cancel_subscriptions
+  StreamSubscription<Position>? _positionStreamSubscription;
+  final Map<Position, Placemark?> _pointList = Map<Position, Placemark?>();
 
   double get latitude => _lastPos.latitude;
   double get longitude => _lastPos.longitude;
@@ -52,12 +53,13 @@ class LocStateModel with ChangeNotifier {
     if (_positionStreamSubscription == null) {
       return true;
     }
-    return _positionStreamSubscription.isPaused;
+    return _positionStreamSubscription!.isPaused;
   }
 
   List<Position> get positions => List.from(_pointList.keys);
   List<Placemark> get placemarks => List.from(_pointList.values);
-  LinkedHashMap<Position, Placemark> get pointList => _pointList;
+  LinkedHashMap<Position, Placemark?> get pointList =>
+      _pointList as LinkedHashMap<Position, Placemark?>;
 
   LocStateModel(this.locator) {
     if (this.locator == null) {
@@ -66,7 +68,7 @@ class LocStateModel with ChangeNotifier {
   }
 
   Future<bool> available() async {
-    _geoAccessStatus = await locator.checkPermission();
+    _geoAccessStatus = await locator!.checkPermission();
     if (_geoAccessStatus == LocationPermission.whileInUse ||
         _geoAccessStatus == LocationPermission.always) {
       return true;
@@ -75,16 +77,14 @@ class LocStateModel with ChangeNotifier {
   }
 
   bool isListening() => !(_positionStreamSubscription == null ||
-      _positionStreamSubscription.isPaused);
+      _positionStreamSubscription!.isPaused);
 
   void addLocation(Position pos) {
     _pointList[pos] = null;
     _lastPos = pos;
     try {
-      locator.placemarkFromCoordinates(pos.latitude, pos.longitude).then((pm) {
-        if (pm != null) {
-          _pointList[pos] = pm.first;
-        }
+      locator!.placemarkFromCoordinates(pos.latitude, pos.longitude).then((pm) {
+        _pointList[pos] = pm.first;
         notifyListeners();
       });
     } catch (e) {
@@ -93,17 +93,11 @@ class LocStateModel with ChangeNotifier {
     notifyListeners();
   }
 
-  Placemark getPlacemark(Position pos) {
-    if (pos == null) {
-      return null;
-    }
+  Placemark? getPlacemark(Position pos) {
     return _pointList[pos];
   }
 
   String getAddressString(Position pos) {
-    if (pos == null) {
-      return '';
-    }
     final placemark = _pointList[pos];
     if (placemark == null) {
       return '';
@@ -126,16 +120,16 @@ class LocStateModel with ChangeNotifier {
 
   void toggleListening() {
     if (_positionStreamSubscription == null) {
-      final Stream<Position> positionStream = locator.getPositionStream(
+      final Stream<Position> positionStream = locator!.getPositionStream(
           desiredAccuracy: LocationAccuracy.best, distanceFilter: 10);
       _positionStreamSubscription =
           positionStream.listen((Position position) => addLocation(position));
       //_positionStreamSubscription.pause();
     } else {
-      if (_positionStreamSubscription.isPaused) {
-        _positionStreamSubscription.resume();
+      if (_positionStreamSubscription!.isPaused) {
+        _positionStreamSubscription!.resume();
       } else {
-        _positionStreamSubscription.pause();
+        _positionStreamSubscription!.pause();
       }
     }
     notifyListeners();
@@ -143,10 +137,7 @@ class LocStateModel with ChangeNotifier {
 
   @override
   void dispose() {
-    if (_positionStreamSubscription != null) {
-      _positionStreamSubscription.cancel();
-      _positionStreamSubscription = null;
-    }
+    _positionStreamSubscription?.cancel();
     super.dispose();
   }
 }
