@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,38 +20,60 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'nb';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "appTitle" : MessageLookupByLibrary.simpleMessage("ActingWeb Første App"),
-    "clickToView" : MessageLookupByLibrary.simpleMessage("Trykk for å se..."),
-    "drawerButtomSheetExpires" : MessageLookupByLibrary.simpleMessage("Utløper"),
-    "drawerButtomSheetFCMToken" : MessageLookupByLibrary.simpleMessage("Firebase meldingstoken"),
-    "drawerButtomSheetIdToken" : MessageLookupByLibrary.simpleMessage("Id token"),
-    "drawerButtomSheetUserToken" : MessageLookupByLibrary.simpleMessage("Bruker token"),
-    "drawerEmail" : MessageLookupByLibrary.simpleMessage("testuser@demoserver.io"),
-    "drawerEmptyEmail" : MessageLookupByLibrary.simpleMessage(""),
-    "drawerEmptyName" : MessageLookupByLibrary.simpleMessage("N/A"),
-    "drawerGetUserInfo" : MessageLookupByLibrary.simpleMessage("Hent brukerinformasjon"),
-    "drawerGetUserInfoFailedMsg" : MessageLookupByLibrary.simpleMessage("Forsøk å oppfrisk tokens først!"),
-    "drawerGetUserInfoFailedTitle" : MessageLookupByLibrary.simpleMessage("Henting av brukerinformasjon feilet"),
-    "drawerGetUserInfoResultMsg" : MessageLookupByLibrary.simpleMessage("Brukerinformasjon har blitt oppdatert."),
-    "drawerGetUserInfoResultTitle" : MessageLookupByLibrary.simpleMessage("Brukerinformasjon oppdatert"),
-    "drawerHeaderInitialEmail" : MessageLookupByLibrary.simpleMessage("..."),
-    "drawerHeaderInitialName" : MessageLookupByLibrary.simpleMessage("Klikk for å se detaljer..."),
-    "drawerLocalisation" : MessageLookupByLibrary.simpleMessage("Nåværende lokalisering"),
-    "drawerLocalisationResultMsg" : MessageLookupByLibrary.simpleMessage("Endret til "),
-    "drawerLocalisationResultTitle" : MessageLookupByLibrary.simpleMessage("Lokalisering endret"),
-    "drawerRefreshTokens" : MessageLookupByLibrary.simpleMessage("Oppfrisk tokens"),
-    "drawerRefreshTokensResultMsg" : MessageLookupByLibrary.simpleMessage("Se bruker detaljer for oppdaterte token."),
-    "drawerRefreshTokensResultTitle" : MessageLookupByLibrary.simpleMessage("Tokens oppfrisket"),
-    "drawerUser" : MessageLookupByLibrary.simpleMessage("Test Bruker"),
-    "latitudeLongitude" : MessageLookupByLibrary.simpleMessage("Lat: \$lat, Long: \$long"),
-    "loginButton" : MessageLookupByLibrary.simpleMessage("Logg inn"),
-    "loginLoadEvents" : MessageLookupByLibrary.simpleMessage("Henter meldinger..."),
-    "loginWelcomeText" : MessageLookupByLibrary.simpleMessage("Velkommen til ActingWeb"),
-    "logoutButton" : MessageLookupByLibrary.simpleMessage("Logg ut"),
-    "okButton" : MessageLookupByLibrary.simpleMessage("Ok"),
-    "startListening" : MessageLookupByLibrary.simpleMessage("Start lytting"),
-    "stopListening" : MessageLookupByLibrary.simpleMessage("Stopp lytting"),
-    "unknown" : MessageLookupByLibrary.simpleMessage("Ukjent")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "appTitle":
+            MessageLookupByLibrary.simpleMessage("ActingWeb Første App"),
+        "clickToView":
+            MessageLookupByLibrary.simpleMessage("Trykk for å se..."),
+        "drawerButtomSheetExpires":
+            MessageLookupByLibrary.simpleMessage("Utløper"),
+        "drawerButtomSheetFCMToken":
+            MessageLookupByLibrary.simpleMessage("Firebase meldingstoken"),
+        "drawerButtomSheetIdToken":
+            MessageLookupByLibrary.simpleMessage("Id token"),
+        "drawerButtomSheetUserToken":
+            MessageLookupByLibrary.simpleMessage("Bruker token"),
+        "drawerEmail":
+            MessageLookupByLibrary.simpleMessage("testuser@demoserver.io"),
+        "drawerEmptyEmail": MessageLookupByLibrary.simpleMessage(""),
+        "drawerEmptyName": MessageLookupByLibrary.simpleMessage("N/A"),
+        "drawerGetUserInfo":
+            MessageLookupByLibrary.simpleMessage("Hent brukerinformasjon"),
+        "drawerGetUserInfoFailedMsg": MessageLookupByLibrary.simpleMessage(
+            "Forsøk å oppfrisk tokens først!"),
+        "drawerGetUserInfoFailedTitle": MessageLookupByLibrary.simpleMessage(
+            "Henting av brukerinformasjon feilet"),
+        "drawerGetUserInfoResultMsg": MessageLookupByLibrary.simpleMessage(
+            "Brukerinformasjon har blitt oppdatert."),
+        "drawerGetUserInfoResultTitle":
+            MessageLookupByLibrary.simpleMessage("Brukerinformasjon oppdatert"),
+        "drawerHeaderInitialEmail": MessageLookupByLibrary.simpleMessage("..."),
+        "drawerHeaderInitialName":
+            MessageLookupByLibrary.simpleMessage("Klikk for å se detaljer..."),
+        "drawerLocalisation":
+            MessageLookupByLibrary.simpleMessage("Nåværende lokalisering"),
+        "drawerLocalisationResultMsg":
+            MessageLookupByLibrary.simpleMessage("Endret til "),
+        "drawerLocalisationResultTitle":
+            MessageLookupByLibrary.simpleMessage("Lokalisering endret"),
+        "drawerRefreshTokens":
+            MessageLookupByLibrary.simpleMessage("Oppfrisk tokens"),
+        "drawerRefreshTokensResultMsg": MessageLookupByLibrary.simpleMessage(
+            "Se bruker detaljer for oppdaterte token."),
+        "drawerRefreshTokensResultTitle":
+            MessageLookupByLibrary.simpleMessage("Tokens oppfrisket"),
+        "drawerUser": MessageLookupByLibrary.simpleMessage("Test Bruker"),
+        "latitudeLongitude":
+            MessageLookupByLibrary.simpleMessage("Lat: \$lat, Long: \$long"),
+        "loginButton": MessageLookupByLibrary.simpleMessage("Logg inn"),
+        "loginLoadEvents":
+            MessageLookupByLibrary.simpleMessage("Henter meldinger..."),
+        "loginWelcomeText":
+            MessageLookupByLibrary.simpleMessage("Velkommen til ActingWeb"),
+        "logoutButton": MessageLookupByLibrary.simpleMessage("Logg ut"),
+        "okButton": MessageLookupByLibrary.simpleMessage("Ok"),
+        "startListening": MessageLookupByLibrary.simpleMessage("Start lytting"),
+        "stopListening": MessageLookupByLibrary.simpleMessage("Stopp lytting"),
+        "unknown": MessageLookupByLibrary.simpleMessage("Ukjent")
+      };
 }
